fix : 최대 가격 설정 및 예외 처리
price 를 long 타입으로, 장바구니 총합 1조원 넘으면 예외

+++ 5week/src/main/java/com/example/kakao/domain/cart/Cart.java
@@ -33,10 +33,10 @@ public class Cart {
     private int quantity;
 
     @Column(nullable = false)
-    private int price; // total price
+    private long price; // total price
 
     @Builder
-    public Cart(int id, User user, Option option, int quantity, int price) {
+    public Cart(int id, User user, Option option, int quantity, long price) {
         this.id = id;
         this.user = user;
         this.option = option;
@@ -45,7 +45,7 @@ public class Cart {
     }
 
     // 장바구니 업데이트
-    public void update(int quantity, int price){
+    public void update(int quantity, long price){
         this.quantity = quantity;
         this.price = price;
     }
